remove auth domain test duplication
